Implement unbound filters as deferred expressions
Unlike other tools, LogQL generally prefers unbound filter statements at
the start of the query, rather than at the end. Sigma allows such
expressions to be deferred - but in the default implementation, they are
appended to the end of the query.

This adds a new class to represent deferred unbound expressions and
prepends any field queries with them during the finalisation process.

Also prepends field queries with `logfmt` to ensure the fields exist.

Updated tests to reflect current status and check unbound filters are
correctly implemented.