Implement ruler output for generating alerts
Loki includes a ruler configuration file (in YAML format) that describes
alerts that can be generated when queries match.

This commit converts Sigma signatures into the ruler YAML format so they
can be used to generate alerts when they are detected. It also includes
a new test to validate the conversion.