Implement NOT operator
As LogQL lacks a NOT operator, we have to manually implement it by
inverting comparisons (i.e., = becomes !=, |~ becomes !~) and inverting
the boolean operators (i.e., !(a AND b) is equivalent to !a OR !b).

This needs a lot more testing, especially for more complex queries. But
for now it works on simple queries.