Fix bug with trailing and operators
If an and/or statement contains only DeferredQueryExpressions as
arguments, joiner.join will return an empty string rather than a None or
a DeferredQueryExpression. Unfortunately, later logic in pySigma is only
looking for those two latter cases, and hence will introduce unnecessary
operators for that empty string - leading to malformed queries.

This appears to be a bug in pySigma in itself - for the time being, I
have imported the relevant logic into the backend and overriden the
check to catch the empty string case - going forward this will be
reported upstream to fix.