Create a new test harness for Sigma backends
Most backends will need to convert a wide range of Sigma rules to be
useful. In order to determine how many different rules a backend can
successfully convert, this script allows a backend or rule developer to
run a specific backend over a directory tree of rules (such as found in
SigmaHQ/sigma) and either see the output rules produced by the backend,
collating any errors that are generated during the process, and even
providing sample test logs to check the rules against.

This is currently written specifically for Loki/LogQL and uses the
logcli command to validate the syntax of the rules and that the query
can be executed without error against the relevant sample log file.
Other backends will need to define their own approach to doing this
process.