Use the fields attributes to control query output
Uses the fields list (if present) in a query with the line_format
command to reduce down the output to a more manageable size.

Considered outputting the results in logfmt style - but given the
end-user can usually expand the row in Grafana to see all the fields,
this seems unnecessary.