Update tester to provide a validation result
Rather than just assume that a lack of stderr means a signature was
valid and worked correctly, include a separate valid boolean in the
results that is used to determine that.

At the moment, I simply check whether logcli writes any output to
stdout (which indicates that a detection event took place).