Ensure stream position is reset for a test file
When using the same test file for multiple rules, the file was being
fully read when validating the first rule, then when it came the others
they just got a file with the stream position at the end of the file.
Instead, seek back to the beginning for each rule validation.