Add code to split large rules based when required
LogQL has a fixed query limit (currently 5120 characters). Some Sigma
rules produce queries that are longer than this - many of which involve
a large number of OR'd field searches. This code changes the behaviour
of convert_rule to partition the rule into multiple smaller rules, and
then converts each of these into its own query.

The approach used for spliting the rule is relatively simple - it
estimates how many rules are required, finds the highest OR in the
query parse tree, and equally divides its arguments between the rules
(copying the parse tree above it and all other rule attributes each
time).