Make partitioning more rigorous
The first approach used to partitioning rules into smaller rules led to
a number of queries that exceeded the max query length in practice.
Some of this was down to the estimation being inaccurate, but also due
to Loki expanding the length of some queries.

This new approach is somewhat more rigourous in that it uses the actual
length of the generated query to decide whether to partition the rule or
not, and only partitions queries that exceed the length requirement
(still using 80% of the hardcoded length due to the above expasion).

This approach will lead to long queries being generated "out of order",
but given most rule files contain just a single rule, the impact of this
should be minimal.

Based on my testing, only 10 generated queries (of 2447) now exceed the
maximum query length for Loki.