Generate line filters from field filters
Loki queries are more performant if they contain line filters (aka
unbound expressions in pySigma terminology). Unfortunately the majority
of Sigma rules do not contain unbound expressions, and hence the
expected performance of the queries is low.

This introduces code that, given a query that initially has no line
filters, attempts to infer from the field filters a valid line filter
that will maximise the length of the line filter, without affecting the
number of results it produces. The assumption is that the longer the
filter, the more effective it will be at reducing the number of
irrelevant log entries that are parsed.

The code currently generates just a single line filter and only produces
exact (negated) matches. Multiple line filters and using regular
expressions may allow for greater coverage of the Sigma rules.