Implement sysmon parsing and overriding log parser
Windows sysmon data of standard promtail output requires significantly
more parsing than just simply using json. To achieve this, the queries
need additional code (including calling a second logfmt parser).

This adds in this functionality, as a separate pipeline, along with the
code to allow custom parsers to be specified in a pipeline.