Fix issue #13 and add regression test
Directly negated ANDs were not being grouped in queries correctly due
to the precendence level being compared between the ConditionNOT and the
parent condition.

This adds in a "peek ahead" that looks ahead to the only argument for
the inner if the inner class is a Condition NOT (which only has a single
argument).