Aggregrate queries in ruler alert format
When producing alerts in Loki, log queries need to produce a
vector/scalar result from an aggregation, rather than the raw log
entries.

To achieve this, wrap each query with:
```
sum(count_over_time(... [1m])) or vector(0) > 0
```
The choice of a 1 minute window was somewhat arbitrary - it is inline
with the default alert query window used in Loki.

Also change the name of the longer alert information from message (which
was used in the Loki documentation) to the more standard description
(used in the Prometheus documentation and expected by Grafana).