Remove outer wildcards from unbound line filters
In Loki, line filters automatically behave as if they start and end with
zero+ length widlcards (.*). Therefore, any such leading or trailing
wildcards in unbound regular expressions can be removed without
affecting the accuracy of the query, and likely improving the overall
query performance.

This also behaves as a workaround for issue grafana/loki#7837