Fix wildcarded unbound OR line filters
The code to convert unbound string expressions did not take into account
the possibility that they might contain wildcards, and hence they were
being passed through to the query without modification (beyond being
escaped as RE metacharacters).

This code fixes that by testing for whether there are any wildcards and
ensuring they are converted appropriately. In the future, this could
also allow full regular expressions from the query to be included;
however, given the wide range of possible inputs and the limited use of
such a feature in Sigma rules, this is still not permitted by this code
and will still lead to an error.