Update case sensitivity for SigmaStrings
SigmaString values should be treated as case insensitive. But Loki
treats equality as being case sensitive. To resolve this, the backend
will now convert all SigmaStrings into regular expressions with the
case insensitive flag set (?i).

Such queries will be less performant - hence the introduction of a
"case_insensitive" boolean flag (default: True). If the performance of a
query is important and the rule is believed to accurately match the case
of the log data, setting this to False will mean the generated query
uses string equality instead (i.e., the query will be case sensitive).