Update tests to reflect new case behaviour
Largely involved changing a=`b` to a=~`(?i)b`, but a few more involved
changes (especially including the line filter functionality, which does
not currently handle case-insensitive searches as effectively as it
could).

Added new test files that were basically identical to the previous
files, but with the case_insensitivity flag set to False to ensure both
cases are covered.