Ensure non-wildcarded string patterns are anchored
LogQL does not anchor regular expressions by default. I.e., the line
filter `| level =~ '.*fo'` will match both `level=info` and
`level=for-real`.

However, we convert strings with wildcards (which expect full value
matching behavior) and hence need to insert those anchors when there
is not a matching wildcard at the start/end of the value.

We may wish to approximate this behavior for `|re` rules in the future,
but it would involve arbitrarily editing user-regexes, which may
introduce other issues.