Avoid adding extraneous pipes to log parsers
On occasion, users may wish to introduce pipes into the start or end of
their log parser configuration. The query finalization would introduce
additional pipes to such a query and cause it to be invalid.

This checks whether or not a log parser already starts or ends with a
pipe and avoids adding one if so.

Closes #156.